requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0
openmeteo_requests
requests-cache
pandas
retry-requests
